// packages/tapable-js/src/tapable.d.ts
type FixedSizeArray<T extends number, U> = T extends 0
  ? void[]
  : ReadonlyArray<U> & {
      0: U;
      length: T;
    };
type Measure<T extends number> = T extends 0 | 1 | 2 | 3 | 4 | 5 | 6 | 7 | 8
  ? T
  : never;
type Append<T extends any[], U> = {
  0: [U];
  1: [T[0], U];
  2: [T[0], T[1], U];
  3: [T[0], T[1], T[2], U];
  4: [T[0], T[1], T[2], T[3], U];
  5: [T[0], T[1], T[2], T[3], T[4], U];
  6: [T[0], T[1], T[2], T[3], T[4], T[5], U];
  7: [T[0], T[1], T[2], T[3], T[4], T[5], T[6], U];
  8: [T[0], T[1], T[2], T[3], T[4], T[5], T[6], T[7], U];
}[Measure<T["length"]>];
type AsArray<T> = T extends any[] ? T : [T];

declare class UnsetAdditionalOptions {
  _UnsetAdditionalOptions: true;
}
type IfSet<X> = X extends UnsetAdditionalOptions ? {} : X;

type Callback<E, T> = (error: E | null, result?: T) => void;
type InnerCallback<E, T> = (error?: E | null | false, result?: T) => void;

type FullTap = Tap & {
  type: "sync" | "async" | "promise";
  fn: Function;
};

type Tap = TapOptions & {
  name: string;
};

type TapOptions = {
  before?: string;
  stage?: number;
};

interface HookInterceptor<T, R, AdditionalOptions = UnsetAdditionalOptions> {
  name?: string;
  tap?: (tap: FullTap & IfSet<AdditionalOptions>) => void;
  call?: (...args: any[]) => void;
  loop?: (...args: any[]) => void;
  error?: (err: Error) => void;
  result?: (result: R) => void;
  done?: () => void;
  register?: (
    tap: FullTap & IfSet<AdditionalOptions>,
  ) => FullTap & IfSet<AdditionalOptions>;
}

type ArgumentNames<T extends any[]> = FixedSizeArray<T["length"], string>;

declare class Hook<T, R, AdditionalOptions = UnsetAdditionalOptions> {
  constructor(args?: ArgumentNames<AsArray<T>>, name?: string);
  name: string | undefined;
  taps: FullTap[];
  intercept(interceptor: HookInterceptor<T, R, AdditionalOptions>): void;
  isUsed(): boolean;
  callAsync(...args: Append<AsArray<T>, Callback<Error, R>>): void;
  promise(...args: AsArray<T>): Promise<R>;
  tap(
    options: string | (Tap & IfSet<AdditionalOptions>),
    fn: (...args: AsArray<T>) => R,
  ): void;
  withOptions(
    options: TapOptions & IfSet<AdditionalOptions>,
  ): Omit<this, "call" | "callAsync" | "promise">;
}